+++ core/src/luigi-config/communication.js
@@ -13,7 +13,7 @@ import { reloadAuth } from './auth/auth';
 import { setFeatureToggle } from './utils/feature-toggles';
 import { setTheme } from './utils/theme';
 import { setSSOAuthData } from './auth/sso';
-import { communicationEntry as pageSizeCommunicationEntry } from './settings/page-size';
+import { communicationEntry as pageSizeCommunicationEntry } from './settings/pagination';
 
 export const communication = {
   customMessagesListeners: {
