+++ core/src/luigi-config/settings/pagination.js
